requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
atomicwrites==1.4.1
attrs==21.4.0
backoff==2.2.1
backports.zoneinfo;python_version<"3.9"
base58==2.1.1
billiard==3.6.4.0
binance==0.3
bitarray==2.6.0
black==22.6.0
boto==2.49.0
botocore==1.28.3
Brotli==1.0.9
CacheControl==0.12.11
cachetools==5.2.0
celery==5.2.7
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
ConfigArgParse==1.5.3
coverage==6.4.2
cytoolz==0.11.2
dateparser==1.0.0
Deprecated==1.2.13
distlib==0.3.5
Django==4.0.6
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-timezone-field==5.0
djangorestframework==3.13.1
dydx-v3-python==1.7.0
ecdsa==0.16.0
eth-abi==2.2.0
eth-account==0.5.8
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
filelock==3.7.1
firebase-admin==5.2.0
frozenlist==1.3.0
gevent==22.10.1
geventhttpclient==2.0.8
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
gql==3.4.0
graphql-core==3.2.3
greenlet==1.1.3.post0
grpcio==1.47.0
grpcio-status==1.47.0
hexbytes==0.2.2
httplib2==0.20.4
hvac==1.0.2
idna==3.3
importlib-metadata
importlib-resources==5.9.0
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.7.2
kombu==5.2.4
locust==2.12.2
lru-dict==1.1.8
MarkupSafe==2.1.1
more-itertools==8.13.0
mpmath==1.0.0
msgpack==1.0.4
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
numpy==1.23.3
packaging==21.3
pandas==1.4.4
parsimonious==0.8.1
pathspec==0.9.0
platformdirs==2.5.2
pluggy==0.13.1
prompt-toolkit==3.0.30
proto-plus==1.20.6
protobuf==3.20.1
psutil==5.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
pyhcl==0.4.4
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==4.6.11
python-binance==1.0.16
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.1
redis==4.3.4
regex==2022.7.24
requests==2.28.1
requests-mock==1.6.0
requests-toolbelt==0.10.0
rlp==2.0.1
roundrobin==0.0.4
rsa==4.9
sentry-sdk==1.9.5
six==1.16.0
sqlparse==0.4.2
sympy==1.6
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tox==3.13.2
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
ujson==5.5.0
uritemplate==4.1.1
urllib3==1.26.12
varint==1.0.2
vine==5.0.0
virtualenv==20.15.1
wcwidth==0.2.5
web3==5.30.0
websocket-client==1.4.1
websockets==9.1
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.5.0
